requirements.txt
python-telegram-bot==20.7
anthropic==0.25.0
PyGithub==2.3.0
aiohttp==3.9.3
